+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Fixed
+- Color attribute value parsing.
+- Style attribute parsing.
 
 ## [0.3.0] - 2017-03-06
 ### Added